requirements.txt
streamlit==1.46.1
streamlit-oauth==0.1.14
opencv-python-headless==4.7.0.72
insightface==0.7.3
numpy==1.23.5
pillow==11.2.1
requests==2.32.4
cloudinary==1.44.1
google-auth==2.40.3
pydantic==2.11.7
python-dotenv==1.0.1
easydict==1.13
scikit-learn==1.7.0
